requirements.txt
google-cloud-logging
python-dotenv
